# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs SWAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/starships": {
            "get": {
                "tags": [
                    "Starships"
                ],
                "summary": "Afficher la liste des vaisseaux spatiaux",
                "operationId": "0a0bde13bcf4e91a94d97408b50e15c4",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des starships"
                    }
                }
            },
            "post": {
                "tags": [
                    "Starships"
                ],
                "summary": "Crée un vaisseau spatial",
                "operationId": "4b8e606ab0ff253bce0652a25f68628e",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "création d'un vaisseau spatial"
                    }
                }
            }
        },
        "/api/starships/{id}": {
            "get": {
                "tags": [
                    "Starships"
                ],
                "summary": "Afficher un vaisseau spatial à l'aide de son id",
                "operationId": "746c1efb4ceccedd68fb022f5440f209",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du vaissaux spatial",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne un starship"
                    }
                }
            },
            "put": {
                "tags": [
                    "Starships"
                ],
                "summary": "Modifier un vaisseau spatial",
                "operationId": "a738f207b45d76972796745bb1d36730",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "modification d'un vaisseau spatial"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Starships"
                ],
                "summary": "Supprimer un vaisseau spatial",
                "operationId": "4e042eb72118e76a7121f00963e104c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du vaisseau spatial",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Suppression d'un vaisseau spatial"
                    }
                }
            }
        },
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Afficher la liste des véhicules",
                "operationId": "b659f082b7df23b2049c6cffcb50cd3f",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des véhicules"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Créer un véhicule",
                "operationId": "fcbb14b08761a447e6e4df8f864409ff",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Création d'un véhicule"
                    }
                }
            }
        },
        "/api/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Afficher un véhicule à l'aide de son id",
                "operationId": "4e2ce1b7fef40fd560c9b21476270e92",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Affiche le véhicule ciblé par l'id passé en paramètre"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Modifier un véhicule à l'aide de son id",
                "operationId": "9d3adda0a00b56831ea9d06e550a0aa3",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Modification d'un véhicule"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Supprimer un véhicule à l'aide de son id",
                "operationId": "95ddbe5ff32596e76c82bab67c10267a",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Suppression d'un véhicule"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}